[adone] refactoring * [collection.Queue] add static "from" method * add pretty namespace * [text] move pretty.table, pretty.json to pretty * [util] move humanizeTime, humanizeSize to pretty
