+++ types/adone/glosses/pretty/index.d.ts
@@ -1,0 +1,10 @@
+/// <reference path="./json.d.ts" />
+/// <reference path="./size.d.ts" />
+/// <reference path="./table.d.ts" />
+/// <reference path="./time.d.ts" />
+
+declare namespace adone {
+    namespace pretty {
+        //
+    }
+}
