+++ types/adone/glosses/pretty/json.d.ts
@@ -1,0 +1,16 @@
+declare namespace adone.pretty {
+    namespace I {
+        interface JSONOptions {
+            emptyArrayMsg?: string;
+            keysColor?: string;
+            dashColor?: string;
+            numberColor?: string;
+            stringColor?: string;
+            defaultIndentation?: number;
+            noColor?: boolean;
+            noAlign?: boolean;
+        }
+    }
+
+    function json(data: any, options?: I.JSONOptions, indentation?: number): string;
+}
