+++ types/adone/glosses/pretty/size.d.ts
@@ -1,0 +1,3 @@
+declare namespace adone.pretty {
+    function size(num: number, space?: string): string;
+}
