+++ types/adone/glosses/pretty/time.d.ts
@@ -1,0 +1,11 @@
+declare namespace adone.pretty {
+    namespace I {
+        interface TimeOptions {
+            msDecimalDigits?: number;
+            secDecimalDigits?: number;
+            verbose?: boolean;
+            compact?: boolean;
+        }
+    }
+    function time(ms: number, options?: I.TimeOptions): string;
+}
