+++ types/adone/glosses/utils.d.ts
@@ -86,17 +86,6 @@ declare namespace adone {
 
         function invertObject(source: object, options?: I.KeysOptions): object;
 
-        namespace I {
-            interface HumanizeTimeOptions {
-                msDecimalDigits?: number;
-                secDecimalDigits?: number;
-                verbose?: boolean;
-                compact?: boolean;
-            }
-        }
-        function humanizeTime(ms: number, options?: I.HumanizeTimeOptions): string;
-        function humanizeSize(num: number, space?: string): string;
-
         function parseSize(str: string | number): number | null;
 
         namespace I {
