+++ types/adone/test/glosses/collections/queue.ts
@@ -14,4 +14,15 @@ namespace adoneTests.collection.Queue {
     { const a: boolean = new Queue().full; }
     { const a: number = new Queue().length; }
     { const a: boolean = new Queue().empty; }
+
+    {
+        const q = Queue.from([1, 2, 3]);
+        const a: number | undefined = q.pop();
+    }
+    {
+        const q = Queue.from((function*() {
+            yield 42;
+        })());
+        const a: number | undefined = q.pop();
+    }
 }
