+++ types/adone/test/glosses/pretty/size.ts
@@ -1,0 +1,12 @@
+namespace adoneTests.pretty.size {
+    const {
+        pretty: {
+            size
+        }
+    } = adone;
+
+    let str: string;
+
+    str = size(12345);
+    str = size(12345, "");
+}
