+++ types/adone/tsconfig.json
@@ -16,7 +16,8 @@
         "types": [],
         "noEmit": true,
         "forceConsistentCasingInFileNames": true,
-        "experimentalDecorators": true
+        "experimentalDecorators": true,
+        "esModuleInterop": true
     },
     "files": [
         "adone-tests.ts",
@@ -63,6 +64,10 @@
         "glosses/net/http/client.d.ts",
         "glosses/net/http/index.d.ts",
         "glosses/net/index.d.ts",
+        "glosses/pretty/json.d.ts",
+        "glosses/pretty/size.d.ts",
+        "glosses/pretty/table.d.ts",
+        "glosses/pretty/time.d.ts",
         "glosses/promise.d.ts",
         "glosses/regex.d.ts",
         "glosses/semver.d.ts",
@@ -78,7 +83,6 @@
         "glosses/templating/index.d.ts",
         "glosses/templating/nunjucks.d.ts",
         "glosses/text/index.d.ts",
-        "glosses/text/pretties.d.ts",
         "glosses/text/spinners.d.ts",
         "glosses/text/table.d.ts",
         "glosses/text/unicode.d.ts",
@@ -123,6 +127,10 @@
         "test/glosses/math/simd.ts",
         "test/glosses/meta.ts",
         "test/glosses/net/http/client.ts",
+        "test/glosses/pretty/json.ts",
+        "test/glosses/pretty/size.ts",
+        "test/glosses/pretty/table.ts",
+        "test/glosses/pretty/time.ts",
         "test/glosses/promise.ts",
         "test/glosses/regex.ts",
         "test/glosses/semver.ts",
@@ -136,7 +144,6 @@
         "test/glosses/templating/dot.ts",
         "test/glosses/templating/nunjucks.ts",
         "test/glosses/text/index.ts",
-        "test/glosses/text/pretties.ts",
         "test/glosses/text/spinners.ts",
         "test/glosses/text/table.ts",
         "test/glosses/text/unicode.ts",
